New GUI and more such as TTS everywhere, TTS dependency in Maven fixed and Pie Charts!

+++ src/main/java/database/Position.java
@@ -26,7 +26,7 @@ public class Position {
     }
 
     public String toString() {
-        String out = x+","+y;
+        String out = (int)x + "," + (int)y;
         return out;
     }
